feat(quasar): add render style registry
This allows mod devs to create custom render styles.

The codec retains backwards compatibility with enum-style names, like `BILLBOARD`.
