+++ common/src/main/java/foundry/veil/VeilClient.java
@@ -8,6 +8,7 @@ import foundry.veil.api.client.render.VeilRenderer;
 import foundry.veil.api.client.render.deferred.VeilDeferredRenderer;
 import foundry.veil.api.event.VeilRenderLevelStageEvent;
 import foundry.veil.api.quasar.registry.EmitterShapeRegistry;
+import foundry.veil.api.quasar.registry.RenderStyleRegistry;
 import foundry.veil.impl.client.editor.*;
 import foundry.veil.impl.client.imgui.VeilImGuiImpl;
 import foundry.veil.impl.resource.VeilResourceManagerImpl;
@@ -72,6 +73,7 @@ public class VeilClient {
         RenderTypeLayerRegistry.bootstrap();
         VeilResourceEditorRegistry.bootstrap();
         EmitterShapeRegistry.bootstrap();
+        RenderStyleRegistry.bootstrap();
     }
 
     @ApiStatus.Internal
