+++ common/src/main/java/foundry/veil/api/quasar/emitters/module/update/FaceVelocityModule.java
@@ -3,6 +3,7 @@ package foundry.veil.api.quasar.emitters.module.update;
 import foundry.veil.api.quasar.emitters.module.UpdateParticleModule;
 import foundry.veil.api.quasar.particle.QuasarParticle;
 import foundry.veil.api.quasar.particle.RenderData;
+import foundry.veil.api.quasar.registry.RenderStyleRegistry;
 import net.minecraft.util.Mth;
 import org.joml.Vector3d;
 import org.joml.Vector3f;
@@ -21,7 +22,7 @@ public class FaceVelocityModule implements UpdateParticleModule {
         Vector3f rotation = particle.getRotation();
         rotation.x = (float) Mth.atan2(normalizedMotion.y, Math.sqrt(normalizedMotion.x * normalizedMotion.x + normalizedMotion.z * normalizedMotion.z));
         rotation.y = (float) Mth.atan2(normalizedMotion.x, normalizedMotion.z);
-        if (particle.getData().renderStyle() == RenderData.RenderStyle.BILLBOARD) {
+        if (particle.getData().renderStyle() == RenderStyleRegistry.BILLBOARD.get()) {
             rotation.y += (float) (Math.PI / 2.0);
         }
     }
